# sdks/sdkwork-payment-app-sdk/openapi/sdkwork-payment-app-api.sdkgen.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Payment App API",
    "version": "1.0.0",
    "description": "App/client payment execution and refund contract for commerce.payment.",
    "x-sdkwork-api-authority": "sdkwork-payment-app-api",
    "x-sdkwork-sdk-family": "sdkwork-payment-app-sdk",
    "x-sdkwork-audience": "Authenticated SDKWork payment clients"
  },
  "servers": [
    {
      "url": "http://localhost:18094",
      "description": "Local payment gateway"
    }
  ],
...
  "security": [
    {
      "AuthToken": [],
      "AccessToken": []
    }
  ],
  "paths": {
...
    "/app/v3/api/payments/intents/{paymentIntentId}": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Retrieve a payment intent.",
        "operationId": "payments.intents.retrieve",
        "parameters": [
          {
            "$ref": "#/components/parameters/PaymentIntentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentIntentResponse"
                }
              }
            }
          },
          "default": {
            "description": "SDKWork problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-forbid-credential-headers": false,
        "x-sdkwork-owner": "sdkwork-payment",
        "x-sdkwork-api-authority": "sdkwork-payment-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "paymentIntents",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "commerce.payments.intents.read"
      }
    },
...
    "/app/v3/api/payments/records/{paymentId}": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Retrieve a payment record.",
        "operationId": "payments.records.retrieve",
        "parameters": [
          {
            "$ref": "#/components/parameters/PaymentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRecordResponse"
                }
              }
            }
          },
          "default": {
            "description": "SDKWork problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-forbid-credential-headers": false,
        "x-sdkwork-owner": "sdkwork-payment",
        "x-sdkwork-api-authority": "sdkwork-payment-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "paymentRecords",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "commerce.payments.records.read"
      }
    },
    "/app/v3/api/payments/attempts/{paymentAttemptId}": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Retrieve a payment attempt.",
        "operationId": "payments.attempts.retrieve",
        "parameters": [
          {
            "$ref": "#/components/parameters/PaymentAttemptId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentAttemptResponse"
                }
              }
            }
          },
          "default": {
            "description": "SDKWork problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-forbid-credential-headers": false,
        "x-sdkwork-owner": "sdkwork-payment",
        "x-sdkwork-api-authority": "sdkwork-payment-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "paymentAttempts",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "commerce.payments.attempts.read"
      }
    },
    "/app/v3/api/payments/statistics/summary": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Retrieve the payment statistics summary.",
        "operationId": "payments.statistics.summary.retrieve",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentStatisticsResponse"
                }
              }
            }
          },
          "default": {
            "description": "SDKWork problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-forbid-credential-headers": false,
        "x-sdkwork-owner": "sdkwork-payment",
        "x-sdkwork-api-authority": "sdkwork-payment-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "paymentStatistics",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "commerce.payments.statistics.read",
        "parameters": []
      }
    },
    "/app/v3/api/payments/checkout/{paymentId}": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Retrieve payment checkout data.",
        "operationId": "payments.checkout.retrieve",
        "parameters": [
          {
            "$ref": "#/components/parameters/PaymentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "default": {
            "description": "SDKWork problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-forbid-credential-headers": false,
        "x-sdkwork-owner": "sdkwork-payment",
        "x-sdkwork-api-authority": "sdkwork-payment-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "paymentCheckout",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "commerce.payments.checkout.read"
      }
    },
    "/app/v3/api/payments/status/{paymentId}": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Retrieve payment status.",
        "operationId": "payments.status.retrieve",
        "parameters": [
          {
            "$ref": "#/components/parameters/PaymentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRecordResponse"
                }
              }
            }
          },
          "default": {
            "description": "SDKWork problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-forbid-credential-headers": false,
        "x-sdkwork-owner": "sdkwork-payment",
        "x-sdkwork-api-authority": "sdkwork-payment-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "paymentStatus",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "commerce.payments.status.read"
      }
    },
    "/app/v3/api/payments/status/out_trade_no/{outTradeNo}": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Retrieve payment status by provider trade number.",
        "operationId": "payments.status.outTradeNo.retrieve",
        "parameters": [
          {
            "name": "outTradeNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRecordResponse"
                }
              }
            }
          },
          "default": {
            "description": "SDKWork problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-forbid-credential-headers": false,
        "x-sdkwork-owner": "sdkwork-payment",
        "x-sdkwork-api-authority": "sdkwork-payment-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "paymentStatus",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "commerce.payments.status.read"
      }
    },
...
    "/app/v3/api/refunds/{refundId}": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Retrieve a refund.",
        "operationId": "refunds.retrieve",
        "parameters": [
          {
            "name": "refundId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundResponse"
                }
              }
            }
          },
          "default": {
            "description": "SDKWork problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-forbid-credential-headers": false,
        "x-sdkwork-owner": "sdkwork-payment",
        "x-sdkwork-api-authority": "sdkwork-payment-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "refunds",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "commerce.refunds.read"
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    },
    "parameters": {
...
      "PaymentId": {
        "name": "paymentId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "PaymentIntentId": {
        "name": "paymentIntentId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "PaymentAttemptId": {
        "name": "paymentAttemptId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
...
    "schemas": {
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "const": 0
          },
          "data": {
            "type": "object"
          },
          "traceId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32",
            "minimum": 40001,
            "maximum": 79999
          },
          "traceId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "PaymentIntent": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "paymentIntentId",
          "orderId",
          "paymentIntentNo",
          "paymentMethod",
          "providerCode",
          "amount",
          "currencyCode",
          "status"
        ],
        "properties": {
          "paymentIntentId": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "paymentIntentNo": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "providerCode": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "currencyCode": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "PaymentRecord": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "paymentId",
          "orderId",
          "outTradeNo",
          "paymentMethod",
          "amount",
          "createdAt",
          "status",
          "statusName"
        ],
        "properties": {
          "paymentId": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "outTradeNo": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "statusName": {
            "type": "string"
          }
        }
      },
...
      "Refund": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "refundId",
          "refundNo",
          "orderId",
          "amount",
          "currencyCode",
          "status"
        ],
        "properties": {
          "refundId": {
            "type": "string"
          },
          "refundNo": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "paymentAttemptId": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "currencyCode": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "reasonCode": {
            "type": "string"
          }
        }
      },
      "PaymentIntentResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "item"
                ],
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/PaymentIntent"
                  }
                }
              }
            }
          }
        ]
      },
      "PaymentAttemptResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "item"
                ],
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/PaymentAttempt"
                  }
                }
              }
            }
          }
        ]
      },
      "PaymentResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "item"
                ],
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/Payment"
                  }
                }
              }
            }
          }
        ]
      },
      "PaymentRecordResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "item"
                ],
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/PaymentRecord"
                  }
                }
              }
            }
          }
        ]
      },
      "PaymentStatisticsResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "item"
                ],
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/PaymentStatistics"
                  }
                }
              }
            }
          }
        ]
      },
      "RefundResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "item"
                ],
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/Refund"
                  }
                }
              }
            }
          }
        ]
      },
...
    }
  }
}